+++ src/main/java/com/examle/sikmogilbackend/record/Calendar/api/dto/CalendarDTO.java
@@ -2,14 +2,13 @@ package com.examle.sikmogilbackend.record.Calendar.api.dto;
 
 import com.examle.sikmogilbackend.record.WorkoutLog.api.dto.WorkoutListDTO;
 import com.examle.sikmogilbackend.record.dietLog.api.dto.DietPictureDTO;
-import lombok.Builder;
-
 import java.util.List;
+import lombok.Builder;
 
 @Builder
 public record CalendarDTO(
         String diaryDate,
-        Long diaryWeight,
+        Double diaryWeight,
         String diaryText,
         List<DietPictureDTO> dietPictures,
         List<WorkoutListDTO> workoutLists
